Refactor to address comments
