+++ fdbrpc/genericactors.actor.h
@@ -418,15 +418,6 @@ Future<X> reportEndpointFailure(Future<X> value, Endpoint endpoint) {
 	}
 }
 
-// Enables connection failures, i.e., clogging, in simulation
-void enableConnectionFailures(std::string const& context);
-
-// Disables connection failures, i.e., clogging, in simulation
-void disableConnectionFailures(std::string const& context);
-
-// Disables connection failures after the giving time seconds
-Future<Void> disableConnectionFailuresAfter(double const& seconds, std::string const& context);
-
 #include "flow/unactorcompiler.h"
 
 #endif
